+++ Functions/Get-PSAddress.ps1
@@ -1,0 +1,36 @@
+<#
+.SYNOPSIS
+    Returns any URLs from application or web configuration.
+.DESCRIPTION
+    It accepts configuration XMLs and returns any URLs found in appSettings and
+    from client endpoint addresses.
+
+    The cmdlet filters PSAppSettings for URLs and also returns PSEndpoint results.
+
+.PARAMETER ConfigXml
+    Mandatory - Pipeline input for Configuration XML
+
+.EXAMPLE
+    Get-PSWebConfig -Path 'C:\inetpub\wwwroot\myapp' | Get-PSAddress
+.EXAMPLE
+    Get-WebSite mysite | Get-PSWebConfig | Get-PSAddress
+#>
+function Get-PSAddress {
+    [CmdletBinding()]
+    param(
+        [Parameter(Mandatory=$true,ValueFromPipeLine=$true)]
+        [psobject[]]$ConfigXml
+    )
+
+    process {
+        # Return all service endpoint addresses
+        Get-PSEndpoint -ConfigXml $configXml
+
+        # Return any URL from appSettings as an Address
+        Get-PSAppSetting -ConfigXml $configXml |
+            Where-Object value -imatch '^http[s]*:' |
+            Add-Member -MemberType AliasProperty -Name name -Value key -Force -PassThru |
+            Add-Member -MemberType AliasProperty -Name address -Value value -Force -PassThru |
+            Set_Type -TypeName 'PSWebConfig.Address'
+    }
+}