+++ Functions/Get-PSAppSetting.ps1
@@ -24,7 +24,7 @@ function Get-PSAppSetting {
         foreach ($config in $ConfigXml) {
             if ($config -is [string]) { $config = [xml]$config }
 
-            if ($config | gm -Name configuration) {
+            if ($config | Get-Member -Name configuration) {
                 if ($config.configuration.appSettings.EncryptedData) {
                     Write-Warning "appSettings section is encrypted. You may not see all relevant entries."
                 }