+++ Tests/Module.tests.ps1
@@ -1,10 +1,14 @@
 . (Join-Path $PSScriptRoot Import-LocalModule.ps1)
 
 $moduleName = 'PSWebConfig'
+$exportedCommands = (Get-Command -Module $moduleName)
 $expectedCommands = @(
     'Get-PSWebConfig'
     'Get-PSAppSetting'
     'Get-PSConnectionString'
+    'Get-PSEndpoint'
+    'Get-PSAddress'
+    
     'Test-PSConnectionString'
 )
 
@@ -14,18 +18,33 @@ Describe "$moduleName Module" {
     }
 }
 
-Foreach ($command in (Get-Command -Module $moduleName))
-{
-    Describe "$moduleName\$command Command" {
-        It "Should be expected" {
-            $expectedCommands -contains $command.Name | Should Be $true
-        }
+Describe 'Exported commands' {
+    # Test if the exported command is expected
+    Foreach ($command in $exportedCommands)
+    {
+        Context $command {
+            It 'Should be an expected command' {
+                $expectedCommands -contains $command.Name | Should Be $true
+            }
 
-        It "Should have proper help" {
-            $help = Get-Help $command.Name
-            $help.description | Should Not BeNullOrEmpty
-            $help.Synopsis | Should Not BeNullOrEmpty
-            $help.examples | Should Not BeNullOrEmpty
+            It 'Should have proper help' {
+                $help = Get-Help $command.Name
+                $help.description | Should Not BeNullOrEmpty
+                $help.Synopsis | Should Not BeNullOrEmpty
+                $help.examples | Should Not BeNullOrEmpty
+            }
+        }
+    }
+}
+
+Describe 'Expected commands' {
+    # Test if the expected command is exported
+    Foreach ($command in $expectedCommands)
+    {
+        Context $command {
+            It 'Should be an exported command' {
+                $exportedCommands.Name -contains $command | Should Be $true
+            }
         }
     }
 }