+++ CHANGELOG.md
@@ -9,10 +9,17 @@
  - Add PowerShell views for WebConfig, ConnectionString objects
  - Add warning for non-admin users for decrypt attemps
 
-## 1.1.0 (Feb 20, 2016)
+## 1.2.0 (Feb 20, 2016)
 - Add views for all `PSWebConfig` object-types
 
 ## 1.3.0 (Feb 25, 2016)
  - Automatic decryption of all configuration sections
  - Add `FileInfo` InputObject support
  - Set Path as the firts positinal parameter
+
+## 1.4.0 (Feb 26, 2016)
+ - Introduce `Get-PSEndpoint` and `Get-PSAddress` to get URLs from config files
+ - Change `FileName` output to `FileInfo`
+ - Add `PSScriptAnalyzer` module for Pester tests
+ - Fix code suggestions from `PSScriptAnalyzer`
+ - Test if expected commands are exported too